// src/utils/globalClasses.d.ts
export declare const proseGroups: {
    base: string;
    text: string;
    largeLight: string;
    image: string;
    quote: string;
    header: string;
};
export declare const proseStyles: {
    base: string;
};
export declare const footerStyles: {
    wave: string;
};
export declare const layoutSpacing: {
    section: string;
    containerWhite: string;
    containerGrey: string;
};
export declare const globalSpace: {
    margin: string;
    padding: string;
    marginX: string;
    marginNegX: string;
    marginY: string;
    paddingY: string;
    paddingYHalf: string;
    paddingX: string;
};
//# sourceMappingURL=globalClasses.d.ts.map